rescaffold database
