+++ FightsApi.sln
@@ -5,13 +5,13 @@ VisualStudioVersion = 16.0.31624.102
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FightsApi", "FightsApi\FightsApi.csproj", "{4C5D08C6-7570-4BB1-BA68-B6342DE8B805}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FightsApi_Data", "FightsApi_Data\FightsApi_Data.csproj", "{A68588C5-ACFF-4A2F-AFC4-7A930087EFB8}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FightsApi_Buisiness", "FightsApi_Buisiness\FightsApi_Buisiness.csproj", "{560878D5-61CD-493B-961B-AA63FC973EBC}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FightsApi_Test", "FightsApi_Test\FightsApi_Test.csproj", "{5CA8250E-8037-43FA-9BE0-8678C62C45C7}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FightsApi_Models", "FightsApi_Models\FightsApi_Models.csproj", "{868F2577-2DEB-45C2-B45E-48D91DD0E3ED}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FightsApi_DbContext", "FightsApi_DbContext\FightsApi_DbContext.csproj", "{CBB2F3F7-C72C-4782-A39C-BA5440755D00}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,10 +23,6 @@ Global
 		{4C5D08C6-7570-4BB1-BA68-B6342DE8B805}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4C5D08C6-7570-4BB1-BA68-B6342DE8B805}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4C5D08C6-7570-4BB1-BA68-B6342DE8B805}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A68588C5-ACFF-4A2F-AFC4-7A930087EFB8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A68588C5-ACFF-4A2F-AFC4-7A930087EFB8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A68588C5-ACFF-4A2F-AFC4-7A930087EFB8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A68588C5-ACFF-4A2F-AFC4-7A930087EFB8}.Release|Any CPU.Build.0 = Release|Any CPU
 		{560878D5-61CD-493B-961B-AA63FC973EBC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{560878D5-61CD-493B-961B-AA63FC973EBC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{560878D5-61CD-493B-961B-AA63FC973EBC}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -39,6 +35,10 @@ Global
 		{868F2577-2DEB-45C2-B45E-48D91DD0E3ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{868F2577-2DEB-45C2-B45E-48D91DD0E3ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{868F2577-2DEB-45C2-B45E-48D91DD0E3ED}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CBB2F3F7-C72C-4782-A39C-BA5440755D00}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CBB2F3F7-C72C-4782-A39C-BA5440755D00}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CBB2F3F7-C72C-4782-A39C-BA5440755D00}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CBB2F3F7-C72C-4782-A39C-BA5440755D00}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
