+++ FightsApi/FightsApi.csproj
@@ -5,11 +5,16 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="5.0.10">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
     <PackageReference Include="Swashbuckle.AspNetCore" Version="5.6.3" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\FightsApi_Buisiness\FightsApi_Buisiness.csproj" />
+    <ProjectReference Include="..\FightsApi_DbContext\FightsApi_DbContext.csproj" />
   </ItemGroup>
 
 </Project>
