+++ FightsApi_Buisiness/Mappers/FighterMapper.cs
@@ -1,5 +1,5 @@
 ﻿using FightsApi_Buisiness.Interfaces;
-using FightsApi_Data;
+using FightsApi_DbContext;
 using FightsApi_Models.ViewModels;
 using System;
 using System.Collections.Generic;
