+++ FightsApi_DbContext/Fight.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 #nullable disable
 
-namespace FightsApi_Data
+namespace FightsApi_DbContext
 {
     public partial class Fight
     {
@@ -18,6 +18,7 @@ namespace FightsApi_Data
         public DateTime? EndDate { get; set; }
         public int? Location { get; set; }
         public int? Weather { get; set; }
+        public bool? Public { get; set; }
 
         public virtual Location LocationNavigation { get; set; }
         public virtual Weather WeatherNavigation { get; set; }
