+++ FightsApi_DbContext/Location.cs
@@ -3,17 +3,17 @@ using System.Collections.Generic;
 
 #nullable disable
 
-namespace FightsApi_Data
+namespace FightsApi_DbContext
 {
-    public partial class Weather
+    public partial class Location
     {
-        public Weather()
+        public Location()
         {
             Fights = new HashSet<Fight>();
         }
 
-        public int WeatherId { get; set; }
-        public string Description { get; set; }
+        public int LocationId { get; set; }
+        public string Location1 { get; set; }
 
         public virtual ICollection<Fight> Fights { get; set; }
     }
