+++ FightsApi_Test/FightsApi_Test.csproj
@@ -20,10 +20,18 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\FightsApi\FightsApi.csproj" />
-    <ProjectReference Include="..\FightsApi_Buisiness\FightsApi_Buisiness.csproj" />
-    <ProjectReference Include="..\FightsApi_Data\FightsApi_Data.csproj" />
-    <ProjectReference Include="..\FightsApi_Models\FightsApi_Models.csproj" />
+    <ProjectReference Include="..\FightsApi\FightsApi.csproj">
+      <TreatAsUsed>true</TreatAsUsed>
+    </ProjectReference>
+    <ProjectReference Include="..\FightsApi_Buisiness\FightsApi_Buisiness.csproj">
+      <TreatAsUsed>true</TreatAsUsed>
+    </ProjectReference>
+    <ProjectReference Include="..\FightsApi_DbContext\FightsApi_DbContext.csproj">
+      <TreatAsUsed>true</TreatAsUsed>
+    </ProjectReference>
+    <ProjectReference Include="..\FightsApi_Models\FightsApi_Models.csproj">
+      <TreatAsUsed>true</TreatAsUsed>
+    </ProjectReference>
   </ItemGroup>
 
 </Project>
